% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mBIC.R
\name{mBIC}
\alias{mBIC}
\title{Modified Bayesian Information Criterion}
\usage{
mBIC(beta, Y, X)
}
\arguments{
\item{beta}{the estimated coefficients.}

\item{Y}{the response.}

\item{X}{design matrix with the same order of the columns in \code{beta}.}
}
\value{
Returns an object with

\item{BIC}{a numeric value with the corresponding BIC.}

\item{K}{the corresponding number of groups.}
}
\description{
Calculate the modified Bayesian information criterion for estimated model.
}
\references{
{'Pairwise Fusion Approach Incorporating Prior Constraint Information' by Yaguang Li}
}
\seealso{
BIC
}
